frontend/requirements.txt
kivy==2.3.1
kivymd==1.2.0
requests==2.31.0
websocket-client==1.6.4